Move EditorPane related preferences(-loading) into EditorPane
